+++ src/lib.rs
@@ -1,1 +1,7 @@
-
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum Version<'a> {
+    Major,
+    Minor,
+    Patch,
+    Custom(&'a str),
+}