+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Usage instructions when no arguments are supplied.
+- Command line argument handling.
+- Usage instructions for `--help` and invalid arguments.
 
 ## [0.1.0] - 2022-10-02